# apps/companion/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Open Resume Companion API",
    "version": "0.1.0",
    "description": "Local companion service for extracting job details from pasted job URLs."
  },
  "components": {
    "schemas": {
...
      "CreateJobRequestInput": {
        "type": "object",
        "properties": {
          "sourceUrl": {
            "type": "string",
            "format": "uri",
            "description": "HTTP or HTTPS job posting URL to crawl."
          }
        },
        "required": [
          "sourceUrl"
        ],
        "additionalProperties": false
      },
...
      "CompanionErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "CrawlStatus": {
        "type": "string",
        "enum": [
          "pending",
          "crawling",
          "analyzing",
          "ready",
          "failed"
        ]
      },
...
      "CompanionJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "crawlStatus": {
            "$ref": "#/components/schemas/CrawlStatus"
          },
          "crawlError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "cleanedText": {
            "type": "string"
          },
          "createdAt": {
            "type": "number",
            "description": "Unix timestamp in milliseconds."
          },
          "updatedAt": {
            "type": "number",
            "description": "Unix timestamp in milliseconds."
          },
          "crawledAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix timestamp in milliseconds, or null before crawl success."
          },
          "parsedTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "parsedCompany": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "parsedLocation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "parsedDescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "fitScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "fitBriefJson": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "sourceUrl",
          "crawlStatus",
          "crawlError",
          "cleanedText",
          "createdAt",
          "updatedAt",
          "crawledAt"
        ],
        "additionalProperties": false
      },
      "CompanionJobsResponse": {
        "type": "object",
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanionJob"
            }
          }
        },
        "required": [
          "jobs"
        ],
        "additionalProperties": false
      },
      "DeleteJobResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "additionalProperties": false
      },
      "ResumeContent": {
        "type": "object",
        "additionalProperties": {}
      },
...
      "JobFitBrief": {
        "type": "object",
        "properties": {
          "roleSummary": {
            "type": "string"
          },
          "requirements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "strengths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gaps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "risks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nextActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "generatedAt": {
            "type": "number"
          }
        },
        "required": [
          "roleSummary",
          "requirements",
          "keywords",
          "strengths",
          "gaps",
          "risks",
          "nextActions",
          "generatedAt"
        ],
        "additionalProperties": false
      },
      "JobApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "company": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "saved",
              "analyzing",
              "tailoring",
              "applied",
              "interviewing",
              "offer",
              "rejected",
              "archived"
            ]
          },
          "sourceResumeId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "sourceResumeName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "sourceResumeSnapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeContent"
              },
              {
                "type": "null"
              }
            ]
          },
          "tailoredResume": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeContent"
              },
              {
                "type": "null"
              }
            ]
          },
          "fitBrief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobFitBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "resumeEditProposals": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "target": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "section": {
                          "type": "string",
                          "const": "summary"
                        }
                      },
                      "required": [
                        "section"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "section": {
                          "type": "string",
                          "const": "experience"
                        },
                        "itemId": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string",
                          "enum": [
                            "role",
                            "description"
                          ]
                        }
                      },
                      "required": [
                        "section",
                        "itemId",
                        "field"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "section": {
                          "type": "string",
                          "const": "experience"
                        },
                        "itemId": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string",
                          "const": "bullet"
                        },
                        "bulletIndex": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "section",
                        "itemId",
                        "field",
                        "bulletIndex"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "section": {
                          "type": "string",
                          "const": "skills"
                        },
                        "itemId": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string",
                          "const": "items"
                        }
                      },
                      "required": [
                        "section",
                        "itemId",
                        "field"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "section": {
                          "type": "string",
                          "const": "projects"
                        },
                        "itemId": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string",
                          "const": "description"
                        }
                      },
                      "required": [
                        "section",
                        "itemId",
                        "field"
                      ],
                      "additionalProperties": false
                    }
                  ]
                },
                "currentText": {
                  "type": "string"
                },
                "suggestedText": {
                  "type": "string"
                },
                "rationale": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "approved",
                    "rejected",
                    "applied"
                  ]
                },
                "createdAt": {
                  "type": "number"
                },
                "appliedAt": {
                  "type": "number"
                }
              },
              "required": [
                "id",
                "target",
                "currentText",
                "suggestedText",
                "rationale",
                "status",
                "createdAt"
              ],
              "additionalProperties": false
            }
          },
          "coverLetterDraft": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "generatedAt": {
                    "type": "number"
                  },
                  "updatedAt": {
                    "type": "number"
                  }
                },
                "required": [
                  "content",
                  "generatedAt",
                  "updatedAt"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "type": "string"
          },
          "followUpAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "number"
          },
          "updatedAt": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "company",
          "title",
          "location",
          "sourceUrl",
          "description",
          "status",
          "sourceResumeId",
          "sourceResumeName",
          "sourceResumeSnapshot",
          "tailoredResume",
          "fitBrief",
          "resumeEditProposals",
          "coverLetterDraft",
          "notes",
          "followUpAt",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "operationId": "createJob",
        "summary": "Create a companion job and enqueue crawling",
        "tags": [
          "Jobs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequestInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionJob"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listJobs",
        "summary": "List companion jobs",
        "tags": [
          "Jobs"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionJobsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "operationId": "getJob",
        "summary": "Get a companion job",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionJob"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteJob",
        "summary": "Delete a companion job",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/screenshot": {
      "get": {
        "operationId": "getJobScreenshot",
        "summary": "Get a job crawl screenshot",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The captured screenshot PNG image.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/retry-crawl": {
      "post": {
        "operationId": "retryJobCrawl",
        "summary": "Retry crawling a companion job",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionJob"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/retry-analyze": {
      "post": {
        "operationId": "retryJobAnalysis",
        "summary": "Retry analysis for a companion job",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionJob"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/convert": {
      "post": {
        "operationId": "convertJobToApplication",
        "summary": "Convert a companion job to a job application",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplication"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}